Xp System
I have made it possible to gain xp and i have made an XP system that scales using the formula in runescapes. There are still many more bugs regarding the combat system but i have managed to make it possible for the stats to transport between scenes. i also made it possible for each individual hero to pick up stached stats and to save their stats