MANY CHANGES
I have Started to make the pause meny where you see everything that should be seen in a pause menu such as characters, stats, item and more, I have also made it so that you can select and remove members of your party and that will change what characters are in the battle. The stats are also transfered and saved so the health your character att the end of the abttle has it will have till the next battle.